requirements.txt
arxiv
dotenv
google-genai
mcp
asyncio